+++ examples/url.rs
@@ -1,9 +1,5 @@
 //To run this example `cargo run --example url --release`
 
-extern crate dont_disappear;
-extern crate read_input;
-extern crate url;
-
 use read_input::prelude::*;
 use read_input::shortcut::with_description;
 use url::Url;
@@ -11,7 +7,7 @@ use url::Url;
 fn main() {
     println!(
         "You inputted the URL {:#?}",
-        input_new::<Url>()
+        input::<Url>()
             .err_match(with_description)
             .repeat_msg("Please input a URL: ")
             .get()