+++ src/lib.rs
@@ -1,187 +1,227 @@
 //! Go the the [readme](https://crates.io/crates/read_input) file for documentation.
 
+#![deny(clippy::pedantic, missing_docs)]
 // `impl ToString` is better than `&impl ToString`. Clippy is not ready for impl trait.
 #![allow(clippy::needless_pass_by_value)]
 
+mod core;
 pub mod prelude;
 pub mod shortcut;
+mod test_generators;
+#[cfg(test)]
+mod tests;
 
-use std::io;
-use std::io::Write;
-use std::str::FromStr;
-use std::string::ToString;
+use crate::{core::read_input, test_generators::InsideFunc};
+use std::{cmp::PartialOrd, rc::Rc, str::FromStr, string::ToString};
 
 const DEFAULT_ERR: &str = "That value does not pass. Please try again";
 
-struct PromptMsg {
-    msg: String,
-    repeat: bool,
+/// Trait for comman types that store input settings.
+pub trait InputBuild<T: FromStr> {
+    /// Changes or adds a prompt message that gets printed once when input if fetched.
+    fn msg(self, msg: impl ToString) -> Self;
+    /// Changes or adds a prompt message and that is repeated each time input is requested.
+    fn repeat_msg(self, msg: impl ToString) -> Self;
+    /// Changes fallback error message.
+    fn err(self, err: impl ToString) -> Self;
+    /// Adds a validation check on input.
+    fn add_test<F: Fn(&T) -> bool + 'static>(self, test: F) -> Self;
+    /// Adds a validation check on input with a custom error message printed when the test
+    /// fails.
+    fn add_err_test<F>(self, test: F, err: impl ToString) -> Self
+    where
+        F: Fn(&T) -> bool + 'static;
+    /// Removes all validation checks made by `.add_test()`, `.add_err_test()`,
+    /// `.inside()` and `.inside_err()`.
+    fn clear_tests(self) -> Self;
+    /// Used specify custom error messages that depend on the errors produced by `from_str()`.
+    fn err_match<F>(self, err_match: F) -> Self
+    where
+        F: Fn(&T::Err) -> Option<String> + 'static;
+    /// Ensures that input is within a range, array or vector.
+    fn inside<U: InsideFunc<T>>(self, constraint: U) -> Self;
+    /// Ensures that input is within a range, array or vector with a custom error message
+    /// printed when input fails.
+    fn inside_err<U: InsideFunc<T>>(self, constraint: U, err: impl ToString) -> Self;
+    /// Toggles whether a prompt message gets printed once or each time input is requested.
+    fn toggle_msg_repeat(self) -> Self;
 }
 
-impl PromptMsg {
-    fn new() -> Self {
-        Self {
-            msg: String::new(),
-            repeat: false,
-        }
+/// Trait for changing input settings by adding constraints that require `PartialOrd`
+/// on the input type.
+pub trait InputConstraints<T>: InputBuild<T>
+where
+    T: FromStr + PartialOrd + 'static,
+    Self: Sized,
+{
+    /// Sets a minimum input value.
+    fn min(self, min: T) -> Self {
+        self.inside(min..)
     }
-    fn from_str(s: impl ToString) -> Self {
-        Self {
-            msg: s.to_string(),
-            repeat: false,
-        }
+    /// Sets a minimum input value with custom error message.
+    fn min_err(self, min: T, err: impl ToString) -> Self {
+        self.inside_err(min.., err)
     }
-    fn repeat_from_str(s: impl ToString) -> Self {
-        Self {
-            msg: s.to_string(),
-            repeat: true,
-        }
+    /// Sets a maximum input value.
+    fn max(self, max: T) -> Self {
+        self.inside(..=max)
+    }
+    /// Sets a maximum input value with custom error message.
+    fn max_err(self, max: T, err: impl ToString) -> Self {
+        self.inside_err(..=max, err)
+    }
+    /// Sets a minimum and maximum input value.
+    fn min_max(self, min: T, max: T) -> Self {
+        self.inside(min..=max)
+    }
+    /// Sets a minimum and maximum input value with custom error message.
+    fn min_max_err(self, min: T, max: T, err: impl ToString) -> Self {
+        self.inside_err(min..=max, err)
+    }
+    /// Sets a restricted input value.
+    fn not(self, this: T) -> Self {
+        self.add_test(move |x: &T| *x != this)
+    }
+    /// Sets a restricted input value with custom error message.
+    fn not_err(self, this: T, err: impl ToString) -> Self {
+        self.add_err_test(move |x: &T| *x != this, err)
     }
 }
 
-type TestFunc<T> = Fn(&T) -> bool;
+#[derive(Clone)]
+pub(crate) struct Prompt {
+    pub msg: String,
+    pub repeat: bool,
+}
 
-/// `InputBuilder` is a 'builder' used to store the settings that are used to fetch input.
+#[derive(Clone)]
+pub(crate) struct Test<T> {
+    pub func: Rc<Fn(&T) -> bool>,
+    pub err: Option<String>,
+}
+
+/// 'builder' used to store the settings that are used to fetch input.
+///
+/// `.get()` method only takes these settings by reference so can be called multiple times.
+///
+/// This type does not have support for default input value.
 pub struct InputBuilder<T: FromStr> {
-    msg: PromptMsg,
+    msg: Prompt,
     err: String,
-    test: Vec<(Box<TestFunc<T>>, Option<String>)>,
-    err_match: Box<dyn Fn(&T::Err) -> Option<String>>,
-}
-
-pub struct InputBuilderOnce<T: FromStr> {
-    builder: InputBuilder<T>,
-    default: Option<T>,
+    tests: Vec<Test<T>>,
+    err_match: Rc<dyn Fn(&T::Err) -> Option<String>>,
 }
 
 impl<T: FromStr> InputBuilder<T> {
     /// Creates a new instance of `InputBuilder` with default settings.
     pub fn new() -> Self {
-        InputBuilder {
-            msg: PromptMsg::new(),
+        Self {
+            msg: Prompt {
+                msg: String::new(),
+                repeat: false,
+            },
             err: DEFAULT_ERR.to_string(),
-            test: Vec::new(),
-            err_match: Box::new(|_| None),
+            tests: Vec::new(),
+            err_match: Rc::new(|_| None),
         }
     }
-    /// Changes or adds a prompt message. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn msg(self, msg: impl ToString) -> Self {
-        InputBuilder {
-            msg: PromptMsg::from_str(msg),
-            ..self
-        }
+    /// 'gets' the input form the user.
+    pub fn get(&self) -> T {
+        read_input::<T>(&self.msg, &self.err, None, &self.tests, &*self.err_match)
     }
-    /// Changes or adds a prompt message and makes it repeat. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn repeat_msg(self, msg: impl ToString) -> Self {
-        InputBuilder {
-            msg: PromptMsg::repeat_from_str(msg),
-            ..self
-        }
-    }
-    /// Changes fallback error message. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn err(self, err: impl ToString) -> Self {
-        InputBuilder {
-            err: err.to_string(),
-            ..self
-        }
-    }
-    /// Changes or adds a default input value. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
+    /// Changes or adds a default input value.
     pub fn default(self, default: T) -> InputBuilderOnce<T> {
         InputBuilderOnce {
             builder: self,
             default: Some(default),
         }
     }
-    fn test<F: 'static + Fn(&T) -> bool>(self, test: F, err: Option<String>) -> Self {
-        InputBuilder {
-            test: {
-                let mut x = self.test;
-                x.push((Box::new(test), err));
+    // Internal function for adding tests and constraints.
+    fn test_err_opt(self, func: Rc<Fn(&T) -> bool>, err: Option<String>) -> Self {
+        Self {
+            tests: {
+                let mut x = self.tests;
+                x.push(Test { func, err });
                 x
             },
             ..self
         }
     }
-    /// Adds a validation check on input. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn add_test<F: 'static + Fn(&T) -> bool>(self, test: F) -> Self {
-        self.test(test, None)
-    }
-    /// Adds a validation check on input with custom error message. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn add_err_test<F: 'static + Fn(&T) -> bool>(self, test: F, err: impl ToString) -> Self {
-        self.test(test, Some(err.to_string()))
-    }
-    /// Removes all validation checks made by `.add_test()` and `.add_err_test()`.
-    pub fn clear_tests(self) -> Self {
-        InputBuilder {
-            test: Vec::new(),
+}
+
+impl<T: FromStr> InputBuild<T> for InputBuilder<T> {
+    fn msg(self, msg: impl ToString) -> Self {
+        Self {
+            msg: Prompt {
+                msg: msg.to_string(),
+                repeat: false,
+            },
             ..self
         }
     }
-    /// Used specify custom error messages that depend on the errors produced by `from_str()`. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn err_match<F: 'static + Fn(&T::Err) -> Option<String>>(self, err_match: F) -> Self {
-        InputBuilder {
-            err_match: Box::new(err_match),
+    fn repeat_msg(self, msg: impl ToString) -> Self {
+        Self {
+            msg: Prompt {
+                msg: msg.to_string(),
+                repeat: true,
+            },
             ..self
         }
     }
-    /// 'gets' the input form the user. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn get(&self) -> T {
-        read_input::<T>(&self.msg, &self.err, None, &self.test, &*self.err_match)
+    fn err(self, err: impl ToString) -> Self {
+        Self {
+            err: err.to_string(),
+            ..self
+        }
+    }
+
+    fn add_test<F: Fn(&T) -> bool + 'static>(self, test: F) -> Self {
+        self.test_err_opt(Rc::new(test), None)
+    }
+    fn add_err_test<F>(self, test: F, err: impl ToString) -> Self
+    where
+        F: Fn(&T) -> bool + 'static,
+    {
+        self.test_err_opt(Rc::new(test), Some(err.to_string()))
+    }
+    fn clear_tests(self) -> Self {
+        Self {
+            tests: Vec::new(),
+            ..self
+        }
+    }
+    fn err_match<F>(self, err_match: F) -> Self
+    where
+        F: Fn(&T::Err) -> Option<String> + 'static,
+    {
+        Self {
+            err_match: Rc::new(err_match),
+            ..self
+        }
+    }
+    fn inside<U: InsideFunc<T>>(self, constraint: U) -> Self {
+        self.test_err_opt(constraint.contains_func(), None)
+    }
+    fn inside_err<U: InsideFunc<T>>(self, constraint: U, err: impl ToString) -> Self {
+        self.test_err_opt(constraint.contains_func(), Some(err.to_string()))
+    }
+    fn toggle_msg_repeat(self) -> Self {
+        Self {
+            msg: Prompt {
+                repeat: !self.msg.repeat,
+                ..self.msg
+            },
+            ..self
+        }
     }
 }
 
-impl<T: FromStr> InputBuilderOnce<T> {
-    pub fn msg(self, msg: impl ToString) -> Self {
-        Self {
-            builder: self.builder.msg(msg),
-            ..self
-        }
-    }
-    pub fn repeat_msg(self, msg: impl ToString) -> Self {
-        Self {
-            builder: self.builder.repeat_msg(msg),
-            ..self
-        }
-    }
-    pub fn err(self, err: impl ToString) -> Self {
-        Self {
-            builder: self.builder.err(err),
-            ..self
-        }
-    }
-    pub fn add_test<F: 'static + Fn(&T) -> bool>(self, test: F) -> Self {
-        Self {
-            builder: self.builder.add_test(test),
-            ..self
-        }
-    }
-    pub fn add_err_test<F: 'static + Fn(&T) -> bool>(self, test: F, err: impl ToString) -> Self {
-        Self {
-            builder: self.builder.add_err_test(test, err),
-            ..self
-        }
-    }
-    pub fn clear_tests(self) -> Self {
-        Self {
-            builder: self.builder.clear_tests(),
-            ..self
-        }
-    }
-    pub fn err_match<F: 'static + Fn(&T::Err) -> Option<String>>(self, err_match: F) -> Self {
-        Self {
-            builder: self.builder.err_match(err_match),
-            ..self
-        }
-    }
-    pub fn get(self) -> T {
-        read_input::<T>(
-            &self.builder.msg,
-            &self.builder.err,
-            self.default,
-            &self.builder.test,
-            &*self.builder.err_match,
-        )
-    }
+impl<T> InputConstraints<T> for InputBuilder<T>
+where
+    T: FromStr,
+    T: PartialOrd,
+    T: 'static,
+{
 }
 
 impl<T: FromStr> Default for InputBuilder<T> {
@@ -190,69 +230,106 @@ impl<T: FromStr> Default for InputBuilder<T> {
     }
 }
 
-/// Creates a new instance of `InputBuilder` with default settings. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-pub fn input_new<T: FromStr>() -> InputBuilder<T> {
-    InputBuilder::new()
+impl<T: FromStr + Clone> Clone for InputBuilder<T> {
+    fn clone(&self) -> Self {
+        Self {
+            msg: self.msg.clone(),
+            err: self.err.clone(),
+            tests: self.tests.clone(),
+            err_match: self.err_match.clone(),
+        }
+    }
 }
 
-fn try_flush() {
-    io::stdout().flush().unwrap_or(())
-}
-
-fn input_str() -> String {
-    let mut input = String::new();
-    io::stdin()
-        .read_line(&mut input)
-        .expect("Failed to read line");
-    input
-}
-
-fn read_input<T: FromStr>(
-    prompt: &PromptMsg,
-    err: &str,
+/// 'builder' used to store the settings that are used to fetch input.
+///
+/// `.get()` method takes ownership of the settings so can be called only once without cloning.
+///
+/// This type has support for default input value.
+pub struct InputBuilderOnce<T: FromStr> {
+    builder: InputBuilder<T>,
     default: Option<T>,
-    test: &[(Box<TestFunc<T>>, Option<String>)],
-    err_pass: &dyn Fn(&T::Err) -> Option<String>,
-) -> T {
-    print!("{}", prompt.msg);
-    try_flush();
+}
 
-    let mut input = input_str();
-
-    if input.trim().is_empty() {
-        if let Some(x) = default {
-            return x;
+impl<T: FromStr> InputBuilderOnce<T> {
+    /// 'gets' the input form the user.
+    pub fn get(self) -> T {
+        read_input::<T>(
+            &self.builder.msg,
+            &self.builder.err,
+            self.default,
+            &self.builder.tests,
+            &*self.builder.err_match,
+        )
+    }
+    // Function that makes it less verbose to change settings of internal `InputBuilder`.
+    fn internal<F>(self, with: F) -> Self
+    where
+        F: FnOnce(InputBuilder<T>) -> InputBuilder<T>,
+    {
+        Self {
+            builder: with(self.builder),
+            ..self
         }
-    };
+    }
+}
 
-    loop {
-        match T::from_str(&input.trim()) {
-            Ok(value) => {
-                let mut test_err = None;
-                let passes_test = test.iter().all(|f| {
-                    if f.0(&value) {
-                        true
-                    } else {
-                        test_err = Some(f.1.clone().unwrap_or_else(|| err.to_string()));
-                        false
-                    }
-                });
-                if passes_test {
-                    return value;
-                } else {
-                    println!("{}", test_err.unwrap_or_else(|| err.to_string()));
-                }
-            }
-            Err(error) => {
-                println!("{}", (err_pass(&error)).unwrap_or_else(|| err.to_string()));
-            }
+impl<T: FromStr> InputBuild<T> for InputBuilderOnce<T> {
+    fn msg(self, msg: impl ToString) -> Self {
+        self.internal(|x| x.msg(msg))
+    }
+    fn repeat_msg(self, msg: impl ToString) -> Self {
+        self.internal(|x| x.repeat_msg(msg))
+    }
+    fn err(self, err: impl ToString) -> Self {
+        self.internal(|x| x.err(err))
+    }
+    fn add_test<F: Fn(&T) -> bool + 'static>(self, test: F) -> Self {
+        self.internal(|x| x.add_test(test))
+    }
+    fn add_err_test<F>(self, test: F, err: impl ToString) -> Self
+    where
+        F: Fn(&T) -> bool + 'static,
+    {
+        self.internal(|x| x.add_err_test(test, err))
+    }
+    fn clear_tests(self) -> Self {
+        self.internal(|x| x.clear_tests())
+    }
+    fn err_match<F>(self, err_match: F) -> Self
+    where
+        F: Fn(&T::Err) -> Option<String> + 'static,
+    {
+        self.internal(|x| x.err_match(err_match))
+    }
+    fn inside<U: InsideFunc<T>>(self, constraint: U) -> Self {
+        self.internal(|x| x.inside(constraint))
+    }
+    fn inside_err<U: InsideFunc<T>>(self, constraint: U, err: impl ToString) -> Self {
+        self.internal(|x| x.inside_err(constraint, err))
+    }
+    fn toggle_msg_repeat(self) -> Self {
+        self.internal(|x| x.toggle_msg_repeat())
+    }
+}
+
+impl<T> InputConstraints<T> for InputBuilderOnce<T>
+where
+    T: FromStr,
+    T: PartialOrd,
+    T: 'static,
+{
+}
+
+impl<T> Clone for InputBuilderOnce<T>
+where
+    T: Clone,
+    T: FromStr,
+{
+    fn clone(&self) -> Self {
+        Self {
+            default: self.default.clone(),
+            builder: self.builder.clone(),
         }
-
-        if prompt.repeat {
-            print!("{}", prompt.msg);
-            try_flush();
-        };
-
-        input = input_str();
     }
 }