+++ src/prelude.rs
@@ -1,1 +1,3 @@
-pub use crate::input_new;
+//! Module to be imported in the style `use read_input::prelude::*;` to get the essential functions and traits.
+
+pub use crate::{shortcut::input, InputBuild, InputConstraints};